test(quic): move QuicStream tests to mirrored test_stream.py
The eight TestQuicStream* classes tested QuicStream, which lives in
quic/stream.py, while sitting in test_connection.py.
This violated the strict test-mirroring law and left stream.py with
no mirrored test file.

Relocate those classes verbatim into a new mirrored test_stream.py,
along with the quic_stream fixture they need and a minimal mock_protocol
fixture.
The mock_protocol fixture stays duplicated in test_connection.py because
the connection tests still use it.
Drop the now-unused quic_stream fixture and QuicStreamResetError import
from test_connection.py, and narrow its module docstring to connection
management.

Test logic is unchanged.
